refactord to add a screens folder to hold the home component
