+++ src/app/app.routes.ts
@@ -1,5 +1,5 @@
-import {Routes} from '@angular/router';
-import {HomeComponent} from './home/home.component';
+import { Routes } from '@angular/router';
+import { HomeComponent } from './screens/home/home.component';
 export const routes: Routes = [
   {
     path: '',
